// SGUEES-SPA/src/app/pages/Shop/com-orden-compra/models/com-orden-compra.ts
export interface ComOrdenCompra {
	CORR_EMPRESA: number;
	ANIO_PERIODO: number;
	CORR_CUADRO_COMPARATIVO: number;
	FECHA_ORDEN_COMPRA: Date;
	NUMERO_ORDEN_COMPRA: Number;
  NOMBRE_PROVEEDOR: string;
	DIRECCION_PROVEEDOR: string;
	NOMBRE_COMERCIAL: string;
	ACTIVIDAD_ECONOMICA: string;
	NOMBRE_CONTACTO: String;
	CORREO_ELECTRONICO_PROVEEDOR: string;
	TELEFONO_PROVEEDOR: string;
	CONDICION_PAGO: String;
  NUMERO_SOLI_COTIZACION: string;
	NOMBRE_DEPARTAMENTO: string;
  OBSERVACIONES: string;
  SUB_TOTAL: Number;
  TOTAL_IMPUESTO: Number;
  TOTAL_ORDEN: Number;
}